fix (#147, #150, #152): Reworked OT update process to resolve core OT trans bugs
This is a commit part of a larger fix and does not fully fix the core issues (yet)
