+++ solardoc/frontend/src/services/phoenix/ot-trans.ts
@@ -1,3 +1,7 @@
+import { useCurrentFileStore } from '@/stores/current-file'
+import { useEditorUpdateWSClient } from '@/stores/editor-update-ws-client'
+import {v4 as uuidv4} from "uuid";
+
 /**
  * A raw operational transformation which inserts text.
  */
@@ -54,4 +58,63 @@ export interface OTrans {
   user_id: string
   acknowledged: boolean
   init: boolean
+}
+
+let enabled = false
+const currentFileStore = useCurrentFileStore()
+const editorUpdateWSClient = useEditorUpdateWSClient()
+
+async function handleIncomingUpdate(toProcess: OTransRespDto): Promise<void> {
+  currentFileStore.pushOTransResp(toProcess)
+}
+
+export async function handleOutgoingUpdate(toSend: OTransReqDto, sendToRemote: boolean = false): Promise<void> {
+  currentFileStore.pushOTransReq(toSend)
+  if (sendToRemote) {
+    await editorUpdateWSClient.wsClient?.sendOTrans(
+      toSend,
+      () => console.log('[handle-ot.ts] Successfully sent OT request to remote'),
+      resp =>
+        console.error(`[handle-ot.ts] Received error response in response to OT request: ${resp}`),
+    )
+  }
+}
+
+/**
+ * Starts the process which will handle the incoming and outgoing OT updates.
+ * @since 0.5.0
+ */
+export function handleOTUpdates() {
+  if (enabled) {
+    return
+  }
+
+  editorUpdateWSClient.wsClient?.listenForOTrans(handleIncomingUpdate)
+  enabled = true
+}
+
+/**
+ * Creates an OTrans object which represents a change to the content.
+ * @param insertOrDeleteTrans The raw insert or delete OTrans.
+ * @returns The OTrans object.
+ * @since 0.7.0
+ */
+function createOTrans(insertOrDeleteTrans: RawInsertOTrans | RawDeleteOTrans): OTransReqDto {
+  return {
+    id: uuidv4(),
+    trans: insertOrDeleteTrans,
+  }
+}
+
+/**
+ * Creates an OTrans object which represents a deletion of text.
+ * @param pos The position at which the text should be deleted.
+ * @param length The length of the text to delete.
+ */
+export function createDeleteOTrans(pos: number, length: number): OTransReqDto {
+  return createOTrans({ type: 'delete', pos, length })
+}
+
+export function createInsertOTrans(pos: number, content: string): OTransReqDto {
+  return createOTrans({ type: 'insert', pos, content })
 }
